requirements.txt
openai-whisper
google-genai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
torch>=2.0.0
tqdm
pydub>=0.25.1
groq>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
